Refactor into modular package structure
