+++ emailbot/__init__.py
@@ -1,6 +1,15 @@
 """Helpers for the email bot."""
 
-from .utils import load_env, setup_logging
+from .utils import load_env, setup_logging, log_error
 from .smtp_client import SmtpClient
+from . import extraction, messaging, bot_handlers
 
-__all__ = ["load_env", "setup_logging", "SmtpClient"]
+__all__ = [
+    "load_env",
+    "setup_logging",
+    "log_error",
+    "SmtpClient",
+    "extraction",
+    "messaging",
+    "bot_handlers",
+]
