+++ emailbot/utils.py
@@ -1,5 +1,7 @@
 import logging
 from pathlib import Path
+from datetime import datetime
+
 from dotenv import load_dotenv
 
 
@@ -22,3 +24,17 @@ def setup_logging(log_file: Path) -> None:
             logging.FileHandler(log_file, encoding="utf-8"),
         ],
     )
+
+
+logger = logging.getLogger(__name__)
+
+
+def log_error(msg: str) -> None:
+    """Log an error message and append it to ``bot_errors.log``."""
+    logger.error(msg)
+    try:
+        err_file = Path(__file__).resolve().parent / "bot_errors.log"
+        with err_file.open("a", encoding="utf-8") as f:
+            f.write(f"{datetime.now().isoformat()} {msg}\n")
+    except Exception:
+        pass
